+++ core/moengage_flutter_android/android/src/main/kotlin/com/moengage/flutter/Constants.kt
@@ -44,3 +44,7 @@ const val METHOD_NAME_PUSH_PERMISSION_PERMISSION_COUNT = "updatePushPermissionRe
 // Asset Location for config.json under moengage_flutter package.
 const val ASSET_CONFIG_FILE_PATH = "flutter_assets/packages/moengage_flutter/config.json"
 const val VERSION_KEY = "version"
+
+// Delete User
+const val METHOD_NAME_DELETE_USER = "deleteUser"
+const val ERROR_CODE_DELETE_USER = "DELETE_USER_ERROR"