+++ core/moengage_flutter_platform_interface/lib/src/moengage_flutter_platform_interface.dart
@@ -8,6 +8,7 @@ import '../src/model/permission_type.dart';
 import '../src/model/properties.dart';
 import '../src/model/push/moe_push_service.dart';
 import 'internal/method_channel_moengage_flutter.dart';
+import 'model/user_deletion_data.dart';
 
 /// Platform Interface for MoEngage Flutter Plugin
 abstract class MoEngageFlutterPlatform extends PlatformInterface {
@@ -201,4 +202,11 @@ abstract class MoEngageFlutterPlatform extends PlatformInterface {
   /// [requestCount] This count will be incremented to existing value
   /// [appId] - MoEngage App ID
   void updatePushPermissionRequestCountAndroid(int requestCount, String appId);
+
+  /// Delete User Data from MoEngage Server
+  /// [appId] - MoEngage App ID
+  /// @returns - Instance of [Future] of type [UserDeletionData]
+  /// @since 1.1.0
+  Future<UserDeletionData> deleteUser(String appId) =>
+      throw UnimplementedError('deleteUser() not implemented for Platform');
 }